src/main/java/com/tj/theater/model/Reservation.java
package com.tj.theater.model;

import static com.tj.theater.util.NumberUtil.roundToTwoDecimals;

public class Reservation {
    private final Customer customer;
    private final Showing showing;
    private final int audienceCount;

    public Reservation(Customer customer, Showing showing, int audienceCount) {
        this.customer = customer;
        this.showing = showing;
        this.audienceCount = audienceCount;
    }

    public double totalFee() {
        return roundToTwoDecimals(showing.getMovieFee() * audienceCount);
    }

    public Customer getCustomer() {
        return customer;
    }

    public Showing getShowing() {
        return showing;
    }

    public int getAudienceCount() {
        return audienceCount;
    }
}